Remove default context param
In GTD system, context is relevant for next actions only and might not
be part of every searches to let the user find notes in which he didn't
set the context. So I removed both the addition of the default context
(note template) and the filter on it in each search done. I also removed
the GtdContext command.